[Do not merge] Learning session
This commit shows how to add a new store, retrieve data form the
server and display it. It's part of the hand-over training of the
project by its main maintainer to the rest of the team.

The next step is to click in the team and display its detail in a new
page.

This PR can be refered to later as an example on how to extend Nakadi
UI functionality.