Display a confirmation dialog when updating schemas
Users often change schemas without fully understading the consequences. This leads to multiple incidents and increased support effort in order to work around and amend such mistakes.

We consider that displaying a message asking for confirmation and alerting users to the potentially dangerous actions they are about to perform could have a positive impact.