Nakadi SQL query output event type monitoring link
When clicking the monitoring link for an event type that is the output
of a Nakadi SQL, users were redirected to a dashboard that wasn't
displaying any data. That's because for such event types a different
dashboard would be better suited, as the source of data has to be a
different one (not Nakadi but Nakadi SQL).