Add read_from option to query creation
When creating a new query, Nakadi SQL would process the entire backlog
of events sitting in the input event types. This lead to a spike in
load and the production of potentially a lot of events.

The decision to process all the events was made when we only had
queries for which this is the only sensible choice, namely FULL OUTER
JOINs.

In a full outer join, not having the complete state of the input event
types makes little sense.

But since then, a lot has changed and currently the most common type
of query are stateless queries, meaning filters and projections. For a
significant amount of users, producing events retroactively is not the
expected behaviour and is the most cost inneficient, thus this field
is introduced with default to 'end'.

There is a chance users might pick 'end' without first understanding
its implications or simply overlooking and later on having their
queries not producing events. But parsing the query to identify if the
input event type is a log compacted event type would be significantly
more complex so we trust users to make a sensible choice.